chore: :building_construction: Reorganización de estructura de DTOs en carpetas modulares
