+++ src/auth/auth.controller.ts
@@ -1,11 +1,11 @@
 import { Body, Controller, Post } from '@nestjs/common'
 import { Public } from 'src/decorators/public.decorator'
 import { AuthService } from './auth.service'
-import { SignInDto } from './dto/sign-in.dto'
+import { SignInDto } from './dto/request/signIn-auth-request.dto'
 import { ApiBadRequestResponse, ApiBearerAuth, ApiBody, ApiOkResponse, ApiOperation } from '@nestjs/swagger'
-import { ForgotPasswordResponseDto } from './dto/forgot-password-response.dto'
-import { SignInResponseDto } from './dto/signIn-response.dto'
-import { ForgotPasswordBodyDto } from './dto/forgot-password-body.dto'
+import { ForgotPasswordResponseDto } from './dto/response/forgotPassword-auth-response.dto'
+import { SignInResponseDto } from './dto/response/signIn-auth-response.dto'
+import { ForgotPasswordBodyDto } from './dto/request/forgotPassword-auth-request.dto'
 import { ApiAuthResponses } from 'src/decorators/apiAuthResponse.decorator'
 import { credentials } from 'amqplib'
 
@@ -32,9 +32,15 @@ export class AuthController {
   @ApiBearerAuth('JWT-auth')
   @Post('forgotpassword')
   @ApiOperation({ summary: 'Request a password reset' })
-  @ApiOkResponse({ description: 'Password reset email sent', type: ForgotPasswordResponseDto })
+  @ApiOkResponse({
+    description: 'Password reset email sent',
+    type: ForgotPasswordResponseDto
+  })
   @ApiAuthResponses()
-  @ApiBody({ description: 'User email', type: ForgotPasswordBodyDto })
+  @ApiBody({
+    description: 'User email',
+    type: ForgotPasswordBodyDto
+  })
   async forgotPassword(@Body('email') email: string) {
     return this.authService.handleForgotPassword(email)
   }
