+++ src/decorators/apiAuthResponse.decorator.ts
@@ -3,7 +3,10 @@ import { ApiBadRequestResponse, ApiUnauthorizedResponse } from '@nestjs/swagger'
 
 export function ApiAuthResponses() {
   return applyDecorators(
-    ApiBadRequestResponse({ description: 'Missing or invalid credentials' }),
+    ApiBadRequestResponse({
+      description: 'Missing or invalid credentials',
+      type: ''
+    }),
     ApiUnauthorizedResponse({ description: 'JWT is missing, invalid or expired' })
   )
 }
