+++ src/patient/dto/request/update-user-request.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types'
+import { CreatePatientDto } from './create-patient-request.dto'
+
+export class UpdatePatientDto extends PartialType(CreatePatientDto) {}
