--- src/patient/dto/update-user.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreatePatientDto } from './create-patient.dto';
-
-export class UpdatePatientDto extends PartialType(CreatePatientDto) {}
