+++ src/user/dto/request/create-user-request.dto.ts
@@ -1,6 +1,6 @@
 import { IsEmail, IsEnum, IsNotEmpty, IsOptional, IsString, IsStrongPassword, ValidateNested } from 'class-validator'
-import { USER_ROLES, UserRole } from '../enums/user-role.enum'
-import { CreatePatientDto } from 'src/patient/dto/create-patient.dto'
+import { USER_ROLES, UserRole } from '../../enums/user-role.enum'
+import { CreatePatientDto } from 'src/patient/dto/request/create-patient-request.dto'
 import { Type } from 'class-transformer'
 import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger'
