+++ src/user/dto/request/update-user-request.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types'
+import { CreateUserDto } from './create-user-request.dto'
+
+export class UpdateUserDto extends PartialType(CreateUserDto) {}
