+++ src/user/user.service.ts
@@ -1,14 +1,14 @@
 import { Injectable } from '@nestjs/common'
 import { InjectModel } from '@nestjs/mongoose'
 import { Model } from 'mongoose'
-import { CreateUserDto } from './dto/create-user.dto'
-import { UpdateUserDto } from './dto/update-user.dto'
+import { CreateUserDto } from './dto/request/create-user-request.dto'
+import { UpdateUserDto } from './dto/request/update-user-request.dto'
 import { User } from './model/user.schema'
 import { hashPassword } from 'src/utils/utils'
 import { IUser } from './interfaces/user.interface'
 import { PatientService } from 'src/patient/patient.service'
 import { USER_ROLES } from './enums/user-role.enum'
-import { CreatePatientDto } from 'src/patient/dto/create-patient.dto'
+import { CreatePatientDto } from 'src/patient/dto/request/create-patient-request.dto'
 import { EmailService } from 'src/email/email.service'
 
 @Injectable()
